+++ src/router.tsx
@@ -3,11 +3,11 @@ import { Route, Routes } from 'react-router-dom';
 import Loading from './components/common/Loading';
 import Layout from './layouts/Layout';
 import Community from './pages/Community';
-import FindId from './pages/Find/Id';
-import FindPassword from './pages/Find/Password';
 import Home from './pages/Home';
 import Map from './pages/Home/Map';
 import Login from './pages/Login';
+import FindId from './pages/Login/Find/Id';
+import FindPassword from './pages/Login/Find/Password';
 import SignUp from './pages/SignUp';
 import Hasang from './pages/Test/Hasang';
 import Seohyun from './pages/Test/Seohyun';
